# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: Код ошибки, например, 400 или 500
        type: integer
      message:
        description: Сообщение об ошибке
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  models.UpdateUserRoleRequest:
    properties:
      role:
        type: string
    type: object
  models.UpdateUsernameRequest:
    properties:
      username:
        type: string
    type: object
  models.User:
    properties:
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  models.UserInfoResponse:
    properties:
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
    url: http://example.com/support
  description: API для интернет-магазина спортивного питания. Содержит функционал
    для управления пользователями, продуктами, заказами и отзывами.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Sports Nutrition Store API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Возвращает данные всех пользователей.
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка всех пользователей
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Позволяет администратору удалить только пользователя с ролью "user"
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно удален
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Некорректные данные запроса или удаление невозможно
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление пользователя с ролью "user"
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Возвращает данные конкретного пользователя по его ID.
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение данных пользователя по идентификатору
      tags:
      - users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Позволяет администратору изменить роль пользователя только с "user"
        на "admin"
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления роли
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Роль пользователя обновлена на администратора
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Некорректные данные запроса или обновление роли невозможно
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление роли пользователя на администратора
      tags:
      - users
  /users/me:
    delete:
      consumes:
      - application/json
      description: Позволяет пользователю удалить свою учетную запись. Администраторы
        не могут удалять себя.
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Учетная запись успешно удалена
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Администратор не может удалить себя
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление своей учетной записи
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Получает информацию о текущем пользователе, включая его имя и роль.
        Пароль в ответе не передается.
      parameters:
      - description: Токен пользователя
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный запрос. Информация о пользователе
          schema:
            $ref: '#/definitions/models.UserInfoResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о пользователе
      tags:
      - users
  /users/me/password:
    patch:
      consumes:
      - application/json
      description: Позволяет авторизованному пользователю изменить свой пароль, требуется
        указать старый и новый пароли
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      - description: Данные для обновления пароля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно обновлен
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Пользователь не авторизован или старый пароль указан неверно
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление пароля пользователя
      tags:
      - users
  /users/me/username:
    patch:
      consumes:
      - application/json
      description: Позволяет авторизованному пользователю обновить свое имя
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        type: string
      - description: Данные для обновления имени пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Имя пользователя успешно обновлено
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Имя пользователя уже занято
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление имени пользователя
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...